requirements.txt
pytest
pytest-xdist
allure-pytest
playwright
pytest-playwright
pytest-html
pytest-base-url